NumericExecutionPass: extract ExecutionVisitor into separate file for reuse in SymbolicExecutionPass
