+++ src/CMakeLists.txt
@@ -86,6 +86,7 @@ add_library( ${PROJECT}-cpp OBJECT
   analyze/PropertyResolverPass.cpp
   analyze/PropertyRevisePass.cpp
 
+  execute/ExecutionVisitor.cpp
   execute/NumericExecutionPass.cpp
   execute/HashMapBase.cpp
   execute/ReferenceConstant.cpp
@@ -189,6 +190,7 @@ ecm_generate_headers( ${PROJECT}_EXECUTE_HEADERS_CPP
     CAMELCASE
   HEADER_NAMES
     ChainedHashMap
+    ExecutionVisitor
     Frame
     FunctionState
     HashMapBase
